requirements.txt
discord==2.3.2
pandas==2.1.0
python-decouple